requirements.txt
opencv-python
mediapipe
numpy
pytube
pytubefix
ffpyplayer
Pillow
